removal of type specified edges to simplify from WD api
