+++ src/main/java/edu/ForceDrawnGraphs/models/Edge.java
@@ -2,6 +2,8 @@ package edu.ForceDrawnGraphs.models;
 
 import java.util.Objects;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import edu.ForceDrawnGraphs.wikidata.models.WikiRecSnak;
+import edu.ForceDrawnGraphs.wikidata.models.WikiRecValue.TXT_VAL_TYPE;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class Edge {
@@ -17,9 +19,12 @@ public class Edge {
     //Default requirement for Jackson
   }
 
-  public Edge(String srcVertexQID, String tgtVertexQID) {
+  public Edge(WikiRecSnak snak, String srcVertexQID) {
     this.srcVertexID = srcVertexQID;
-    this.tgtVertexID = tgtVertexQID;
+    this.tgtVertexID = findSnakTargetQID(snak);
+    this.propertyQID = snak.property().value();
+    this.value = findSnakTargetValue(snak);
+    this.datatype = snak.datatype();
   }
 
   public String srcVertexID() {
@@ -33,6 +38,23 @@ public class Edge {
   public void setTgtVertexID(String tgtID) {
     this.tgtVertexID = tgtID;
   }
+
+  public void setPropertyQID(String propertyQID) {
+    this.propertyQID = propertyQID;
+  }
+
+  public String propertyQID() {
+    return propertyQID;
+  }
+
+  public String value() {
+    return value;
+  }
+
+  public String datatype() {
+    return datatype;
+  }
+
 
   @Override
   public boolean equals(Object o) {
@@ -49,4 +71,39 @@ public class Edge {
   public int hashCode() {
     return Objects.hash(srcVertexID, tgtVertexID, weight);
   }
+
+  @Override
+  public String toString() {
+    return "[:propertyQID=" + urlPrefixer(propertyQID) + ", value=" + value + ", datatype="
+        + datatype + ", tgtVertexQID=" + urlPrefixer(tgtVertexID()) + ", srcVertexQID="
+        + urlPrefixer(srcVertexID()) + " :]";
+  }
+
+  //------------------------------------------------------------------------------------------------------------
+  //
+  //! PRIVATE METHODS - PRIVATE METHODS - PRIVATE METHODS - PRIVATE METHODS - PRIVATE METHODS - PRIVATE METHODS
+  //
+  //------------------------------------------------------------------------------------------------------------
+  private String urlPrefixer(String qid) {
+    if (qid == null)
+      return null;
+    if (qid.startsWith("P")) {
+      return "https://www.wikidata.org/wiki/Property:" + qid;
+    } else {
+      return "https://www.wikidata.org/wiki/" + qid;
+    }
+  }
+
+  private static String findSnakTargetQID(WikiRecSnak snak) {
+    if (snak.value().type() != TXT_VAL_TYPE.ENTITY)
+      return null;
+    return snak.value().value();
+  }
+
+  private static String findSnakTargetValue(WikiRecSnak snak) {
+    if (snak.value().type() == TXT_VAL_TYPE.TIME)
+      return snak.value().value();
+    return null;
+  }
+
 }
